fixed the most common memory leak errors that occur when working with byte buffers

+++ src/main/java/com/github/adamorgan/api/requests/Response.java
@@ -36,7 +36,7 @@ public class Response
 
     protected final Exception exception;
 
-    public Response(byte version, byte flags, short stream, byte opcode, int length, ErrorResponse error, ByteBuf body, UUID trace)
+    public Response(byte version, byte flags, int stream, byte opcode, int length, ErrorResponse error, ByteBuf body, UUID trace)
     {
         this.version = version;
         this.flags = flags;
